microservices/gene_edit_service/pyproject.toml
[project]
name = "gene-edit-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=23.2.0",
    "fastapi>=0.104.1",
    "numpy>=1.24.0",
    "orjson>=3.9.0",
    "pandas>=2.1.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "torch>=2.0.0",
    "torch-geometric>=2.4.0",
    "uvicorn[standard]>=0.24.0",
]
